move Trainer class to utils
